+++ cpskin/core/browser/file.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from cpskin.core.utils import safe_utf8
+
 from plone import api
 from plone.app.contenttypes.browser.file import FileView as BaseFileView
 
@@ -9,10 +9,7 @@ class FileView(BaseFileView):
     def __call__(self):
         context = self.context
         if api.user.is_anonymous():
-            download_url = '{0}/@@download/file/{1}'.format(
-                context.absolute_url(),
-                safe_utf8(context.file.filename),
-            )
-            self.request.response.redirect(download_url)
+            pdf_url = context.absolute_url()
+            self.request.response.redirect(pdf_url)
             return ''
         return super(FileView, self).__call__()